requirements.txt
--allow-external PyAudio
--allow-unverified PyAudio
PyAudio==0.2.8
SoundAnalyse==0.1.1
